+++ spec/overcommit/hook_context/pre_commit_spec.rb
@@ -198,19 +198,22 @@ describe Overcommit::HookContext::PreCommit do
       end
     end
 
-    context 'when a broken symlink is staged' do
-      around do |example|
-        repo do
-          Overcommit::Utils::FileUtils.symlink('non-existent-file', 'symlink')
-          `git add symlink`
-          example.run
+    # Git cannot track Windows symlinks
+    unless Overcommit::OS.windows?
+      context 'when a broken symlink is staged' do
+        around do |example|
+          repo do
+            Overcommit::Utils::FileUtils.symlink('non-existent-file', 'symlink')
+            `git add symlink`
+            example.run
+          end
         end
-      end
 
-      it 'does not attempt to update/restore the modification time of the file' do
-        File.should_not_receive(:mtime)
-        File.should_not_receive(:utime)
-        subject
+        it 'does not attempt to update/restore the modification time of the file' do
+          File.should_not_receive(:mtime)
+          File.should_not_receive(:utime)
+          subject
+        end
       end
     end
   end
@@ -512,51 +515,54 @@ describe Overcommit::HookContext::PreCommit do
       end
     end
 
-    context 'when changing a symlink to a directory during an amendment' do
-      around do |example|
-        repo do
-          `git commit --allow-empty -m "Initial commit"`
-          FileUtils.mkdir 'some-directory'
-          symlink('some-directory', 'some-symlink')
-          `git add some-symlink some-directory`
-          `git commit -m "Add file"`
-          `git rm some-symlink`
-          FileUtils.mkdir 'some-symlink'
-          touch File.join('some-symlink', 'another-file')
-          `git add some-symlink`
-          example.run
+    # Git cannot track Windows symlinks
+    unless Overcommit::OS.windows?
+      context 'when changing a symlink to a directory during an amendment' do
+        around do |example|
+          repo do
+            `git commit --allow-empty -m "Initial commit"`
+            FileUtils.mkdir 'some-directory'
+            symlink('some-directory', 'some-symlink')
+            `git add some-symlink some-directory`
+            `git commit -m "Add file"`
+            `git rm some-symlink`
+            FileUtils.mkdir 'some-symlink'
+            touch File.join('some-symlink', 'another-file')
+            `git add some-symlink`
+            example.run
+          end
+        end
+
+        before do
+          context.stub(:amendment?).and_return(true)
+        end
+
+        it 'does not include the directory in the list of modified files' do
+          subject.should_not include File.expand_path('some-symlink')
         end
       end
 
-      before do
-        context.stub(:amendment?).and_return(true)
-      end
-
-      it 'does not include the directory in the list of modified files' do
-        subject.should_not include File.expand_path('some-symlink')
-      end
-    end
-
-    context 'when breaking a symlink during an amendment' do
-      around do |example|
-        repo do
-          `git commit --allow-empty -m "Initial commit"`
-          FileUtils.mkdir 'some-directory'
-          touch File.join('some-directory', 'some-file')
-          symlink('some-directory', 'some-symlink')
-          `git add some-symlink some-directory`
-          `git commit -m "Add file"`
-          `git rm -rf some-directory`
-          example.run
+      context 'when breaking a symlink during an amendment' do
+        around do |example|
+          repo do
+            `git commit --allow-empty -m "Initial commit"`
+            FileUtils.mkdir 'some-directory'
+            touch File.join('some-directory', 'some-file')
+            symlink('some-directory', 'some-symlink')
+            `git add some-symlink some-directory`
+            `git commit -m "Add file"`
+            `git rm -rf some-directory`
+            example.run
+          end
         end
-      end
 
-      before do
-        context.stub(:amendment?).and_return(true)
-      end
+        before do
+          context.stub(:amendment?).and_return(true)
+        end
 
-      it 'still includes the broken symlink in the list of modified files' do
-        subject.should include File.expand_path('some-symlink')
+        it 'still includes the broken symlink in the list of modified files' do
+          subject.should include File.expand_path('some-symlink')
+        end
       end
     end
   end
